+++ frontend/src/App.jsx
@@ -7,28 +7,28 @@ import { Toaster } from "react-hot-toast";
 import "./assets/style/app.scss";
 import "@fontsource-variable/inter";
 
-import DNSAnswers from "./App/component/DNSAnswers";
-import PrivateServers from "./App/PrivateServers";
-import ServerDevices from "./App/ServerDevices";
-import ScreenLoader from "./App/ScreenLoader";
-import InspectGroup from "./App/InspectGroup";
-import UserSelect from "./App/UserSelect";
-import Enable2FA from "./App/Enable2FA";
-import Settings from "./App/Settings";
-import Devices from "./App/Devices";
-import Tunnels from "./App/Tunnels";
-import Account from "./App/Account";
-import Welcome from "./App/Welcome";
-import SideBar from "./App/SideBar";
+import DNSAnswers from "./components/DNSAnswers";
+import PrivateServers from "./pages/PrivateServers";
+import ServerDevices from "./pages/ServerDevices";
+import ScreenLoader from "./pages/ScreenLoader";
+import InspectGroup from "./pages/InspectGroup";
+import UserSelect from "./pages/UserSelect";
+import Enable2FA from "./pages/Enable2FA";
+import Settings from "./pages/Settings";
+import Devices from "./pages/Devices";
+import Tunnels from "./pages/Tunnels";
+import Account from "./pages/Account";
+import Welcome from "./pages/Welcome";
+import SideBar from "./pages/SideBar";
 import GLOBAL_STATE from "./state";
-import Groups from "./App/Groups";
-import Login from "./App/Login";
+import Groups from "./pages/Groups";
+import Login from "./pages/Login";
 import { STATE } from "./state";
-import Users from "./App/Users";
-import Stats from "./App/Stats";
-import Logs from "./App/Logs";
+import Users from "./pages/Users";
+import Stats from "./pages/Stats";
+import Logs from "./pages/Logs";
 import STORE from "./store";
-import DNS from "./App/dns";
+import DNS from "./pages/dns";
 import WS from "./ws";
 
 const appElement = document.getElementById("app");
