+++ frontend/src/components/ConfigDNSRecordEditor.jsx
