+++ frontend/src/components/InfoItem.jsx
