+++ frontend/src/components/ObjectEditorDialog.jsx
