+++ frontend/src/components/interactive-table.jsx
