+++ frontend/src/pages/Account.jsx
@@ -2,11 +2,11 @@ import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import GLOBAL_STATE from "../state";
 import dayjs from "dayjs";
-import KeyValue from "./component/keyvalue";
+import KeyValue from "../components/keyvalue";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import GenericTable from "./GenericTable";
 import { Button } from "@/components/ui/button";
-import InfoItem from "./component/InfoItem";
+import InfoItem from "../components/InfoItem";
 import { Network } from "lucide-react";
 import { Input } from "@/components/ui/input";
 import { DropdownMenuItem } from "@/components/ui/dropdown-menu";
