+++ frontend/src/pages/GenericTable.jsx
