+++ frontend/src/pages/InspectGroup.jsx
