+++ frontend/src/pages/Logs.jsx
