+++ frontend/src/pages/NewObjectEdior.jsx
