+++ frontend/src/pages/NewObjectEdiorDialog.jsx
