+++ frontend/src/pages/ServerDevices.jsx
