+++ frontend/src/pages/Stats.jsx
